Midi: Use span dynamics render results for note velocities. Fixes #1586.
Using Audio_dynamic's volume_ when announced does not work for
span dynamics: only in stop_translation_timestep the actual
volume is rendered and filled-in.

Also, make using note-velocity vs midi dynamics switchable.  Currently
it it always on, as it is mandatory when using the default
midiChannelMapping = 'instrument (as we need different volume settings
in voices that share a channel) and probably preferrable in all
other cases too.