Event listener to extract (some) music events.
This may be useful on its own, or as a basis for other people to
create/modify their own event listeners.

I considered a few different names; I'm not sold on
event-listener.ly.  Other contenders were staff-events.ly,
vivi-events.ly, and vivi-notes.ly.  In the end, I thought that it
would be better to avoid a Vivi-specific name.  Even though this
file is closely tied to Vivi, I don't think it's inappropriate to
include it -- we include the Festival song stuff, after all.