Added new Pagesize definition to paper.scm
Added 'c9landscape' to allow a very small @lilypond to be defined for examples
that need to show footers, taglines and the like without having to

1. Define the papersize within the actual @lilypond; and so make the example more
complicated that it need be for users.

2. Avoid unnecessary white space between the content and footers/taglines etc.

Chosen size is named c9landscape because the new papersize=x option for
the @lilypond [] construct cannot take spaces or use the standard 'landscape
option. Keeping a standard ISO size means we don't have 'custom' sizes in the
scm file.