Fix "make doc" error: Character set messup in pdf-scheme.cc
In "make doc", somehow the LANG env variable is cleared, so
glib returned charset=ANSI_X3.4-1968 for the input string. This would
disallow any accented characters in the input string.
However, we know that the input string is by definition always UTF-8,
so the solution is not to use the current locale, but hardcode UTF-8.

Once we switch to guile 2.0, we should use (sting->utf16 str 'big)
instead of manually converting the string in C.