Separate flags into their own sub-font.
When we have lots of flag sizes, there will be lots of glyphs.

With a maximum of 256-32 = 234 glyphs per sub font, we wanted to
have flags in their own sub font.

But flags depend on notehead widths, so they needed to be in the
same subfont as the noteheads.

This patch moves the notehead parameters from feta-noteheads.mf to
feta-params.mf, so the basic notehead parameters are available to
all subfonts.

Also, AFAICS, bigcheese.pe.in is no longer used -- it doesn't show
up anywhere in a git grep, except as having been renamed.  It doesn't
show up in GNUmakefile.  So I deleted it.