Implements multiple-line non-cross-staff glissandi.
Glissandi, not being placed outside the staff, are not bumped up or
down depending on other outside staff grobs that may appear after
a line break.  Thus, all they need to know are the terminus positions
on the Y axis in order to be typeset.  This means that they can span
multiple lines.