Fix 1214: cueDuring and quoteDuring should also quote voices that create subvoices
If a voice was quoted that created a subvoice, add-quotable wrongly selected
only that subvoice for quoting rather than the original voice.

The proper fix is to correctly use the assoc list returned by
recording-group-emulate. We need to hand it a proper unique voice name
and can then extract the events for just this voice, no matter which other
voices are created.

If the music expression passed to \addQuote is a voice itself, we only use
its contents for quoting.