Fix 1563: System start bars interpreted collapse-height as absolute length.
If you increased the staff-space, this meant sometimes the collapse-height
would not be enough to hide the start-bar for a staff, while in other cases
it was enough...

This patch interprets the collapse-height in multiples of the staff-space.

However, I think that the notion of a collapse-height (as a length) for
hiding/showing the system start delimiter is not the best approach in
general. Sooner or later, we should change the system to show/hide the
system start bar/bracket depending on the number of staves involved rather
than on grob height.