Fix 153: \once\set properly restores the context property
\once\set works by installing a finalization hook in the iterator.
To restore the context property value before the \once\set, we simply
cache the old value and pass it to the finalization hook as third
argument. The finalization hook then sends a SetProperty event with
the old value rather than an UnsetProperty event, which would revert
the value to the default.