Proper loglevels: cmd-line option --loglevel=NONE/ERROR/WARN/BASIC/PROGRESS/DEBUG
Allow the user to specify which messages (s)he wants to see on stderr:
The lilypond code basically stays the same, I only added a loglevel
global variable, which is used in the warning/error*/progress*/success
functions (ly:warning, ly:error, etc. in Scheme). If the proper level
is not set for a message, it is not printed.

There are only some larger changes:
-) Global var be_verbose_global replaced by loglevel (in warn.cc, accessor
   is_loglevel(...);  much finer-grained)
-) New functions debug_output, which replaces code like:
   if (be_verbose_global) {
	   progress_indication (...)
   }
-) Move all scheme log functions to warn-scheme.cc
-) Add (optional) source location info to warning/error/log messages

All changes were done to warn.cc and to the member of the Input class
(apparently, we have two parallel error-reporting "frameworks" in
lilypond...).

Note for all functions in warn.cc (not for the members of Input):
The debug_output function (and progress_indication and message) have
an optional argument to specify whether the output of the message
should always start on a new line or continue the previous output.

All functions of the Input class now are just a front-end for the functions
in warn.cc

Documentation for this new feature is still missing (both in the AU
as well as in the CG).