Doc: NR 5.5.4 - Modifying ties and slurs
Started off as adding a @cindex for 'control points'.

Ended up tidying up the overall description of how the bezier
curve describes its arc.

Then noted a reference to a non-existant property of the TieColumn
object - so I have referred to the correct interface for this property
and added an @seealso.