Fixes issue 620.
Also fixes issue 591 (which was previously fixed, but is re-fixed
here).

Does this by making the extents of an axis group maleable depending
on the interfaces one chooses to use for bounds in the property
bound-alignment-interfaces

See generic_bound_extent in axis-group-interface.cc to see
how this works.  All other changes get the data pumped into
this function.