[doc] Handle leading dash in @code.
Due to the default value of texinfo.tex's `@allowcodebreaks true', line
breaks after a dash within @code is allowed; this makes sense due to the
very long Scheme identifiers containing many dashes.  However, we have a lot
of `@code{-1}', and it looks very ugly to have a line break within this
expression.

This patch fixes it (also converting many `@code' to the correct `@option'),
using `@w{...}' where appropriate.  However, I've done it only for the
English documentation, expecting the translators to follow up.