Guile v2: Set 'show-file-name option correctly.
This fixes a compile issue with Guile v2.

Using the `debug-enable' procedure with non-boolean debug options raises
an error with Guile v2.  The internal type for `show-file-name' is
SCM_OPTION_SCM, and this is why we can't use `debug-enable' here.

The fix is to use `debug-set!', passing the unquoted option name and
the value as arguments (as stated in the Guile documentation).