Lilypond-book: Get rid of lilyquote option, use quote instead
The only difference between the lilyquote and quote options
so far was that with lilyquote the texinfo and doctitle
was outside the example block. We didn't use quote together
with those texinfo texts, so there is really no need to
have both.

Furthermore, those texinfo options were documented as 'obscure'
options, used internally for the lilypond documentation,
so this shouldn't break external scores, either.