Better pure height approximations of stem tremolos.
Before this patch, stem-tremolo pure heights wer always centered at
the offset of their Y parent (often 0), which means that tremolos
were not being considered correctly in horizontal spacing.  This
patch fixes that.  The regression test stem-tremolo-note-column.ly
now correctly checks for this - before, it only worked because flags
were combined with stems and created a Y-extent that overshot the
flag ending. Making more accurate flag extents broke this regression
test, and this patch rectifies it.