Lets auto numbering of footnotes kick in from commands alone.
This patch gets rid of footnote-auto-numbering in the paper
bloc, allowing footnotes to kick in from commands alone.
These commands are:

auto-footnote (for top-level markups)
autoFootnote (for inside chords)
autoFootnoteGrob (for normal grob-footnoting grob situations)

Their homologues remain the same - to wit:

footnote
footnote
footnoteGrob

Currently, if one mixes automatic and non-automatic footnotes, the
automatic footnotes skip over the non-automatic ones.  So,

\auto-footnote foo
\footnote 42 bar
\auto-footnote hello

Will number the items 1, 42, 3: not 1, 42, 2.  The option to make it
1, 42, 2 should be the subject of a future patch.

The problem of get_footnotes_from_lines being called twice still
remains and also should be solved in a separate patch.