Merge segfault regtests
We haven't had a release in a month.  More to the point, we
haven't had a regtest comparison in a month.
http://code.google.com/p/lilypond/issues/detail?id=1821
is not going away any time soon.  It involves complicated stuff in
the build system, which nobody wants to touch.

As a temporary workaround, I will now reduce the number of
regression tests.  This commit merges segfault fixes.