Fix 1821: Write pathes for ly-to-tely to a separate file rather than passing as cmd line args
For out-of-source builds, the collated files list has the whole absolute path
for each and every file (>1000 for our regtests), which blows up the cmd line
length to values that don't work with the shell (<=131000 chars allowed...).

This patch takes the list in the makefile and writes the filenames in several
patches (300 at a time for now) to a .list file and then passes just this one
file to lys-to-tely, which reads in the whole long list of file names from that
file. The trick is that while make can handle long variables, we can't pass
it to ly-to-tely on the cmd line, but have to use a helper file.