Implements unpure-pure-containers in LilyPond.
These structures free functions from needing to be pre-registered as
pure in define-grobs.scm.  For example, if one has a function foo
that is the unpure callback for a given property and bar that is the
pure callback, the generalized syntax for these containers is:

\override Grob #'property = #(ly:make-unpure-pure-container foo bar)

All of the pure/unpure distinctions that require a back and forth
between .scm and .cc files can, in theory, be eliminated by
using these containers.  Furthermore, all of the functions whose
pure equivalents are hard-coded can now take different pure equivalents
(see input/regression/unpure-pure-container.ly).