Flush all ports after each file, flush stdout after displaying all options
So far, the output of the safe.ly regtest showed up randomly
in some later regtest, because stdout was buffered when redirected
to a file.
This patch flushes all ports after each processed file (only relevant
when multiple files are processed at once).

Unfortunately, it seems stdout is not an open Scheme port when
Scheme doesn't print anything to stdout. So, when displaying all
options, we have to flush stdout directly after printing all
options.