Fix 380: Try to auto-detect cyclic references in header fields
Inside \fromproperty, simply change the props so that all references to the
currently interpreted property point to a function that prints out
a warning and returns a null markup, while interpreting the contents
of the property.

This is a specialized case of the cycle detection, since the hare+
tortoise algorithm does not tell us where the cycle was causes.
For properties we have a simple way to detect cyclic dependencies,
so we can tell the user which definition caused the problem.