Build dependencies for .texi files (issue 1852).
Write .dep files containing make dependency rules for .texi and .itexi files.
They allow to simply type `make' to process the doc after changing an included
file and have the manuals updated with the minimum amount of processing.

These .dep files are generated by recursively scanning for lines starting with
@include in the .texi files, and looking up these included files within the
include directories that we pass to texinfo. The same is done for .itexi files.
The .dep files are included into the build by stepmake/generic-targets.make.

With this we can clean up Documentation/GNUmakefile a bit:
  - Remove the previous apparently unsuccessfull attempt at tracking
    dependencies with a wildcard, and
  - Remove dependencies that are caught automatically: only generated files
    need an explicit dependency, which weblinks.itexi is an example, thus
  - Add the explicit dependency for weblinks.itexi.