musicxml2ly: title and subtitle (issue 1913), miscellaneous musicxml2ly: titles (fixes issue 1913), tagline, conversion-info, <source>, midi-cmd-line-option, center-column long instrument names
1) if XML: <work-title>bli</work-title> AND <movement-title>bla
</movement-title> --> LilyPond: title = bla subtitle = bli

2) the tagline of a piece engraved by LilyPond should not contain any
information as to the encoding software of the .xml file. The standard
Lilypond-tagline should be used.

3) the conversion info should contain the name of the conversion tool

4) the <source>-element is converted to a custom LilyPond-variable
named "source" in the header. (it is usually used for publishing
information)

5) a command line option for a midi block was added

6) multi-lined instrument names are now typeset in center-columns

7) the <miscellaneous>-element is currently translated into a header-
variable "texidoc" which is important for the documentation. I'd suggest
to call it "miscellaneous" by default and call it "texinfo" when
activated via a command line option. (Not implemented yet)