Website: use external $LILYPOND_WEB_MEDIA_GIT
This builds the website using the pictures and ly-examples which
will appear on the actual website.  Previously, it used either
pictures from a local "make doc" run, or from a hard-coded
$HOME/lilypond/web-media/ directory.  The latter is how it's built
on lilypond.org, but that relies on me personally uploading new
files.

Once this is accepted and pushed, I'll dig out our "publications"
pdfs, and anything else remaining in web/, add them to this
script, then finally rename the web/ dir (which is the pre-2009
website) to oldweb/.

Building with "make doc" is completely unaffected.