parser.yy: Allow numbers, fractions and \default as arguments.
Common parts of function_arglist and function_arglist_closed are
also factored out in order to avoid premature parser decisions.

In closed music expressions (mostly in the context of optional
arguments), numbers with units (3\cm) and "wide" fractions (3 / 4) are
not recognized, but if the respective number is backed up because of
a false predicate, they can still be used in that manner in a
mandatory argument.

As a consequence of moving most of the optional argument checking into
predicates, the operator priority system could be vastly simplified.

\default can be used to force the rest of an optional argument block
to get skipped.  It is the only way to skip optional arguments if they
are not followed by a mandatory argument, since otherwise a skipped
value could not be interpreted anywhere else in the argument list.