Adds in-notes to LilyPond.
In notes function like footnotes in that they serve to annotate music,
but are different in that they are typeset either above or below
the system to which the grob being annotated belongs.

Currently, in-notes are not implemented in top-level markups.  This will
hopefully be the subject of a future patch.

This patch also does some clean up in page-breaking.cc and
page-layout-problem.cc to speed up the performance of scores with
in-notes and/or footnotes.