Prevents dynamics from colliding with cross-staff stems.
Creates an Interval_minefield class that tracks intersections of
intevals and returns the closest "safe spots" to the original interval
on either side.  Dynamics are then moved to one of these two safe spots
depending on the collision-bias property (with negative values forcing
the decision towards the left and positive towards the right).