Doc: Two doc improvements for LP CLI in Usage
In response to Tracker Items 2114 and 2112

2112
Clarified explanation as suggested by listing the backends and added
additional options to avoid the addition files generated by -dpreview
as suggested.

Changed once instance of back-end to 'backend' as this seems to be
the standard term in the AU doc.

2114
Application Usage 1.2

Changed the way the @table is formatted so that the single quotes are
removed from the options listed.

Took some time to edit some of the linebreaks within the *.itely file so that
it follows the line lengths as per the CG.

Minor line spacing edits to make existing @cindex and @item enries
clearer to see in the source file for future editing