Adds vertical skylines to slurs.
This assures that objects with an outside staff priority higher than
that of slurs are placed directly above the slur at their horizontal
position instead of above the slur's extrema in a given direction.

This is done by giving slurs vertical skylines that are the integral of
their curve.  Because the axis-group-interface automatically uses
vertical-skylines, no extra check for this property has to be added.