Update lilygit.tcl (Issue 2092)
Makes lilygit.tcl respect the environment variable $LILYPOND_GIT.
If $LILYPOND_GIT is unset, default of $HOME/lilypond-git will be used.

Also does all working on dev/local_working branch, instead of master

Adds a Push Patch button to push patch to staging, which is optionally
configured, such that it only displays if push access is available
and the user is not a translations user.

Add support for the environment variable LILYPOND_BRANCH, which can
be prepended to the call to lily-git.tcl to specify the branch to be used.

Add pushHead so we can base patches on master but push to staging

Add log preview so that the user is expected to review the git log before
pushing.

Rebase before pushing.  The rebasing is to origin/staging,
and is performed on a detached head before pushing.  This preserves the
working branch in case of staging breaking and being recreated.

Add environment variable PUSH_ACCESS for controlling push access