Build: Fix for cp: cannot stat `web.texi': No such file or directory.
We have seen an intermittent bug in `make doc' where `web.texi' was
not found when it came time to copy it. This would happen in the cs/
directory first, then running `make doc' again would complete the cs/
translations but die on the de/ or es/ ones. Repeating `make doc'
enough times would complete the doc build.

I had this happen again and after the 3rd `make doc' decided to try
this patch. The next `make doc' completed all of the translations.