Build: Fix lilypond-book regtests when dblatex is not found.
When dblatex is not installed, only the docbook .xml files are
created by a `make doc' but not the .pdf files. We need to filter
those .xml files out of the list of collated files, otherwise
lilypond-book thinks that they are musicxml files and fails to
process them through musicxml2ly.