Set context properties to allow half-measure beaming in 3/4 (Issue 2246)
Introduce beamWholeMeasure, which allows beaming of a whole measure of
eighth notes in 3/4 time (but not beaming of a half measure, since it
appears to be a 6/8 beaming).  By default, set to #t.

Introduce beamHalfMeasure, which allows beaming of a half measure of
eighth notes in 3/4 time.  By default, set to #f.

A non-empty beamExceptions entry will override beamHalfMeasure and
beamWholeMeasure, since user-specified rules should override defaults.

Includes updated documentation in English docs, and a new regtest.  No
changes to translated docs.