rall. and accel stopped working in articulate.
According to the the git logs, in February 2011 the way that tempo
changes for MIDI were interpreted changed from just the
tempoWholesPerMinute property to use a TempoChangeEvent.

This broke rall. in articulated MIDI output.

This patch brings the articulate script up-to-date wrt to the way that
tempo changes are signalled, and also adds accel. and poco accel. to
the list of accepted scripts.

Reported-by: Thomas Morley <thomasmorley65@googlemail.com>
Reviewed-by: Carl Sorensen <Carl.D.Sorensen@gmail.com>
Reviewed-by: Graham Percival <graham@percival-music.ca>
Signed-off-by: Peter Chubb <peter.chubb@nicta.com.au>