Doc: CG - Update of Quick Start section
I refreshed instructions for new contributors for LilyDev and Lily-git.

Main changes include updated instructions on how to install Lily-Dev
in VirtualBox. A complete walk-through on how to download and setup
a LilyDev Virtual Machine.

Added more step-by-step information for Lily-git based on the newer
 version of lily-git.tcl.

Added some more @node and @*sections for consistency and ease of
finding from the Table of Contents.

Made appropriate @ref{} changes in other *.itexi files that point
to this section.