Don't use virtual base classes without apparent reason
Virtual base classes are a tool for joining the representations of common
ancestors in a class lattice.  They come at a cost in computational,
conceptual, and usability complexity.

I can't see the purpose for the instances of virtual inheritance in
LilyPond (without multiple inheritance, there can't be much of an effect).