PO: modifying po-replace before integrating it to the release process
Until now, updating the translation template file is disconnected from
the releasing process, which means that a dev or translator must

1- run 'po-replace',
2- embed lilypond.pot headers,
3- eliminate all po files that have been moved upwards from po/out to
po/
4- commit the new lilypond.pot
5- run 'make dist' to have a tarball,
6- notify the Free Translation Project where to pick up that tarbal.

The first three steps may already be reduced to one single step by
embedding lilypond.pot directly (with specific options and some 'sed'
when creating) and moving it only in po/, what does this patch.

A second step would be to have a call to po-replace when the Release
Meister enters the dance.

We then will have the same unique tarball to refer, directly from
linuxaudio.