Issue 2607: Allow immediate Scheme expressions with splicing and multiple values
$xxx has been able to interpret 0-valued expressions (namely
SCM_UNSPECIFIED) so far by, well, not interpolating anything.

This change allows it to interpret multi-valued expressions by
returning multiple tokens, allowing things like

{ < $(apply values (map (lambda (i) (ly:make-pitch 0 i 0)) (iota 8))) > }

to work as intended, by creating a chord enclosing all 8 pitches.

In addition, in analogy to the unquote-splicing operator ,@ the form

{ < $@(map (lambda (i) (ly:make-pitch 0 i 0)) (iota 8)) > }

is provided for interpolating a list of Scheme expressions into the
token stream.