Reject \footnote ... \default without grob specification
The behavior of \footnote ... \default without grob specification is,
for most purposes, indistinguishable from user error.  In consequence,
it has been turned into user error.

As one consequence, support for the undocumented and inefficient
compatibility mode for use of not-actually-time-based footnotes as an
articulation (-\footnote ... \default) has been removed.  This can
occur as the result of convert-ly from older variants of \footnote
syntax.  Targetting a NO_SMART conversion error message of convert-ly
for this use case is so much more tricky and error-prone than printing
a warning at run-time that it has not been attempted.